+++ packages/test-cases/utils.ts
@@ -1,0 +1,3 @@
+export const debugLog = (message: string, ...data: unknown[] ) => {
+  if (!process.env.CI) console.log(message, ...data)
+}