+++ packages/tx/examples/transactions.ts
@@ -43,18 +43,18 @@ console.log('--------------------')
 // You can decode it with the rlp module.
 // After that you should have something like:
 const rawTx = [
-  '0x00',
+  '0x',
   '0x09184e72a000',
   '0x2710',
   '0x0000000000000000000000000000000000000000',
-  '0x00',
+  '0x',
   '0x7f7465737432000000000000000000000000000000000000000000000000000000600057',
   '0x1c',
   '0x5e1d3a76fbf824220eafc8c79ad578ad2b67d01b0c2425eb1f1347e8f50882ab',
   '0x5bd428537f05f9830e93792f90ea6a3e2d1ee84952dd96edbae9f658f831ab13',
 ]
 
-const tx2 = Transaction.fromValuesArray(rawTx.map(toBuffer)) // This is also a maninnet transaction
+const tx2 = Transaction.fromValuesArray(rawTx.map(toBuffer)) // This is also a mainnet transaction
 
 // Note rlp.decode will produce an array of buffers.
 // So assuming that you were able to parse the transaction, we will now get the sender's address.